Fix compilation for GCC 4.9.2 (compiled using OpenMP)

+++ source/threadingzeug/include/threadingzeug/parallelfor.h
@@ -13,7 +13,7 @@ namespace threadingzeug
 {
 
 
-THREADINGZEUG_API int getNumberOfThreads();
+THREADINGZEUG_API size_t getNumberOfThreads();
 
 
 // Necessary for type deduction with lambdas
