+++ source/threadingzeug/source/parallelfor.cpp
@@ -6,7 +6,7 @@ namespace threadingzeug
 {
 
 
-int getNumberOfThreads()
+size_t getNumberOfThreads()
 {
 	return std::max(static_cast<size_t>(2), static_cast<size_t>(std::thread::hardware_concurrency()));
 }
